Might have fixed something
Former-commit-id: a7e9d770c957f0e6f088e61e8c3ac48457ca4883 [formerly a0d7704d99e9ba1ea913477a76ac8e4989ecb37d]
Former-commit-id: 824a1ce80d1609fa92400714e0e17789a53c4a48

+++ src/librapid/math/mpir.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 // MPIR (modified) for BigNumber types
+#include <cstdint>
 #include <mpirxx.h>
 #include <thread>
 #include <future>
@@ -23,14 +24,14 @@ namespace librapid {
 	class Chudnovsky {
 	public:
 		explicit Chudnovsky(int64_t dig10 = 100);
-		[[nodiscard]] detail::PQT compPQT(int64_t n1, int64_t n2) const;
+		[[nodiscard]] detail::PQT compPQT(int32_t n1, int32_t n2) const;
 		[[nodiscard]] mpf pi() const;
 
 		[[nodiscard]] mpf piMultiThread() const;
 
-		[[nodiscard]] static detail::PQT compPQT2(const Chudnovsky &chud, int64_t n1, int64_t n2,
+		[[nodiscard]] static detail::PQT compPQT2(const Chudnovsky &chud, int32_t n1, int32_t n2,
 												  int64_t depth = 0) {
-			int64_t m;
+			int32_t m;
 			detail::PQT res;
 
 			if (n1 + 1 == n2) {
